Chapter 1 getting started.
NOTE: I did install intellij IDEA rust project, which opened a cargo for
me automatically.

THe mindmap at docs/chapter_01-03-hello-cargo.md picks my big goal,
that is to generate a file-based integration system that integrators
can use, all written in rust.

There are many concerns to be tackled, that are all in the mind map.