Guessing game: random numbers and the crazy crates
In this commit there are my experiments regarding
rust crates.

Even something as trivial as a random number generator is an optional
crate ridden with customizations and options. And it must be declared
in the cargo.toml .

Even more, crates are so, so more complex than your typical java
packaged jar... Inside a single crate you can have multiple
options and choose implementations from inside a use directive.

CRAZY! In a good(ish) sense?? But all this complexity makes
understanding the documentation HELL for the uninitiated.