crates are SOURCE ONLY! match=switch on steroids!
Today it downed on me that CRATES are SOURCE CODE. Not a binary package,
like a jar would be.

Which probably justify why I hear complaints that rust is slow to
compile.

Discovered also about shadowing (no need to crate bla_str,  bla_int ,
bla_etc if the original value is not used).

And the match expression, and a skim on how enums are much more complex
than java enums in Rust.