RUST has shadowing in the same scope!
Today I've studied a little more about rust and the mapping between
java x rust.

I discovered today that the rust constant is quite different than
a static final int x , because constants are defined in COMPILE
TIME ONLY. There is no construction, hence it is not possible
to statically initialize a class with some arbitrary runtime
constant (such as an environment variable, for example).