arrays and tuples - control are axpressions!
In this session I've learned that:

Statements end with semicolon;

Statements don't return a value;

Expressions return stuff, and return a value.

loop and if can be expressions.

while and for _can_ be expressions, but they
can't return anything other than (), so they
are not useful.

For loops only works for "iterable things".

No traditional integer for loop.

To make iterable things, use the range operator, such as  `1..5`

The rust lang documentation has detailed information about
all reserved keywords, and can provide important insight on
what the language can do.