Ch5 - first experiments with structures.
In this chapter I'm creating a table of event structures.

It shows the odds of a single event to happen as a fraction

and shows the min / max values of a natural rand(float 0-1 exclusive)
for that to happen.

At least this is the plan: the code is not compiled yet: I want to log
the evolution of this code.

This is my first attempt on using array::map .