Interesting issue with cargo.toml workspaces.
Today I've stumbled upon an issue with the cargo.toml files.

Rust tries to parse every cargo.toml file in directory hierarchy
to look for workspaces.

A forum post was created to discuss the issue.

The solution was to add an empty src/lib.rs at the main directory.

And also add an [workspace] tag into the ch8_3.