shadowing doesn't drop scope!
Today I've learned that shadowing a variable don't make it to become
out of scope automatically (so it can still be borrowing stuff and
consuming memory).

I've also discovered how to run the application to show assembly.

Discovered also about the existence of a MIR thing, that
deals with borrowing .

I've also progressed a little more in the example of chapter 8_3.